https://doi.org/10.1007/s11276-018-1756-6
This approach assumes all messages are sent in normal mode and analysis only includes speed as the dimension observed. Little detail was given in the paper on how to analyze other data present through BSMs, so I'm using this as a direct comparison to my current algorithm approach. (The code is terrible, but the paper was cryptic in their approach)