Uploaded Recent Changes
Previous code used static values for trust threshold and for speed assessment. New implementation considers a percentile of the reported speed limit for speed assessment, and a percentile of trust value for trust assessment.

Also implemented historical data for trust metric changes - degree of trust at reporting time influences the degree to which the trust score will be updated.